backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
firebase-admin==6.4.0
google-cloud-firestore==2.13.1
google-auth==2.25.2

